Relaxes `request` dependency version.
The newest version of `request` uses patched versions of `hawk` and
`is-my-json-valid`, recently highlighted for DoS vulnerabilities.

Since `request` 2.53 uses `hawk` 2.x, which wasn't patched, the only
way to fix this vulnerability is to update to the most recent `request`
version.

See: https://github.com/hueniverse/hawk/issues/171

This is generally a good practice to use semver dependency versioning.