Handle and report invalid notify(err) argument. Fixes #110.
notify(err…) now catches and reports the case where err=null|undefined|0|true|false.

An error-level log is printed and Error is constructed and sent to Bugsnag
with the offending parameter. All of the arguments are captured and sent
in the report's metadata too, to help provide context for when this happens.