feat(sessions): Enable session tracking by default
This change enables session tracking by default. It also updates the way in which the
notify/sessions endpoint URLs are configured. "useSSL", "notifyHost", "notifyPath", "notifyPort" are
deprecated, as is "sessionEndpoint". The new way to configure endpoint URLs is in the "endpoints"
option – see the docs for more info. Using the deprecated options will still work but a message will
be output to the console.