Fix types for the configuration object (#144)
With 2.4.0 the way in which URLs are configured has been changed. It looks like these types have not been updated in a while so I went through the documentation and updated them. I've added more detailed types to some of the properties that were `any` types. Lastly I've organized them alphabetically.

- `appType`
- `autoCaptureSessions`
- `autoNotifyUnhandledRejection`
- `endpoints`
- `sessionEndpoint`